Made ActionRPG the main project folder
