--- ActionRPG/src/main/PlayRPG.java
@@ -1,26 +1,0 @@
-package main;
-
-
-
-import javax.swing.JFrame;
-
-public class PlayRPG {
-    public static void main(String[] args) throws Exception {
-        JFrame window = new JFrame();
-        window.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        window.setResizable(false);
-        window.setTitle("RPG GAME");
-
-        GamePanel gamePanel = new GamePanel();
-        window.add(gamePanel);
-
-        window.pack();
-
-
-        window.setLocationRelativeTo(null);
-        window.setVisible(true);
-
-        gamePanel.startGameThread();
-        
-    }
-}
