--- ActionRPG/src/res/maps/map01.txt
@@ -1,12 +1,0 @@
-1 1 1 1 1 1 1 1 1 1 1 0 0 0 0 0
-1 1 1 1 1 1 1 1 1 0 0 0 0 0 0 0
-1 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0
-1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
-0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
-3 3 3 3 3 3 3 3 0 0 0 0 0 0 0 0
-2 2 2 2 2 2 2 3 3 3 3 3 0 0 0 0
-2 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3
-2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
-2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
