--- ActionRPG/src/tile/Tile.java
@@ -1,9 +1,0 @@
-package tile;
-
-import java.awt.image.BufferedImage;
-
-public class Tile {
-    public BufferedImage image;
-    public boolean collision = false;
-
-}
