+++ README.md
@@ -1,2 +1,5 @@
 # ACTIONRPG
 APCSA RPG Project
+
+//TODO//
+make character sprites(2 down, 2 left, 2 right, 2 up)
