+++ bin/entity/Entity.class
