+++ bin/entity/Player.class
