+++ bin/res/player/New Piskel-6.png.png
