+++ bin/res/tiles/Stone.png
