+++ bin/tile/TileManager.class
