+++ src/main/GamePanel.java
@@ -1,0 +1,89 @@
+package main;
+
+import javax.swing.JPanel;
+
+import entity.Player;
+import tile.TileManager;
+
+import java.awt.Dimension;
+import java.awt.Graphics;
+import java.awt.Color;
+import java.awt.Graphics2D;
+
+public class GamePanel extends JPanel implements Runnable {
+
+    // Screen Settings
+    final int originalTileSize = 32; // 16x16 pixels
+    final int scale = 2;
+
+    public final int tileSize = originalTileSize * scale; // 48x48 scale
+    public final int maxScreenCol = 16;
+    public final int maxScreenRow = 12;
+    public final int screenWidth = maxScreenCol * tileSize; // 768 pixels
+    public final int screenHeight = maxScreenRow * tileSize; // 576 pixels
+
+    TileManager tileM = new TileManager(this);
+    KeyHandler keyH = new KeyHandler();
+    Thread gameThread;
+    Player player = new Player(this, keyH);
+
+    int FPS = 60;
+
+    public GamePanel() {
+        this.setPreferredSize(new Dimension(screenWidth, screenHeight));
+        this.setBackground(Color.BLACK);
+        this.setDoubleBuffered(true);
+        this.addKeyListener(keyH);
+        this.setFocusable(true);
+
+    }
+
+    public void startGameThread() {
+        gameThread = new Thread(this);
+        gameThread.start();
+    }
+
+    @Override
+    public void run() {
+        double drawInterval = 1000000000 / FPS;
+        double delta = 0;
+        long lastTime = System.nanoTime();
+        long currentTime;
+        long timer = 0;
+        int drawCount = 0;
+
+        while (gameThread != null) {
+            currentTime = System.nanoTime();
+            delta += (currentTime - lastTime) / drawInterval;
+            timer += currentTime - lastTime;
+            lastTime = currentTime;
+
+            if (delta >= 1) {
+                update();
+                repaint();
+                delta--;
+                drawCount++;
+            }
+
+            if (timer >= 1000000000) {
+                System.out.println("FPS: " + drawCount);
+                drawCount = 0;
+                timer = 0;
+            }
+        }
+    }
+
+    public void update() {
+        player.update();
+    }
+
+    public void paintComponent(Graphics g) {
+        super.paintComponent(g);
+
+        Graphics2D g2 = (Graphics2D) g;
+        tileM.draw(g2);
+        player.draw(g2);
+        g2.dispose();
+    }
+
+}
