+++ src/res/tiles/Stone.png
