+++ src/tile/TileManager.java
@@ -1,0 +1,101 @@
+package tile;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+
+import javax.imageio.ImageIO;
+
+import main.GamePanel;
+
+import java.awt.Graphics2D;
+
+public class TileManager {
+
+    GamePanel gp;
+    Tile[] tile;
+    int mapTileNum[][];
+
+    public TileManager(GamePanel gp) {
+        this.gp = gp;
+
+        tile = new Tile[10];
+        mapTileNum = new int[gp.maxScreenCol][gp.maxScreenRow];
+        getTileImage();
+        loadMap("/res/maps/map01.txt");
+    }
+
+    public void getTileImage() {
+        try {
+            tile[0] = new Tile();
+            tile[0].image = ImageIO.read(getClass().getResourceAsStream("/res/tiles/Grass.png")); // tile[0] is grass
+
+            tile[1] = new Tile();
+            tile[1].image = ImageIO.read(getClass().getResourceAsStream("/res/tiles/Stone.png")); // tile[1] is stone
+
+            tile[2] = new Tile();
+            tile[2].image = ImageIO.read(getClass().getResourceAsStream("/res/tiles/Water.png")); // tile[2] is water
+
+            tile[3] = new Tile();
+            tile[3].image = ImageIO.read(getClass().getResourceAsStream("/res/tiles/Sand.png")); // tile[3] is sand
+
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+    }
+
+    public void loadMap(String filePath) {
+        try {
+            InputStream is = getClass().getResourceAsStream(filePath);
+            BufferedReader br = new BufferedReader(new InputStreamReader(is));
+
+            int col = 0;
+            int row = 0;
+
+            while (col < gp.maxScreenCol && row < gp.maxScreenRow) {
+                String line = br.readLine();
+
+                while (col < gp.maxScreenCol) {
+                    String numbers[] = line.split(" ");
+                    int num = Integer.parseInt(numbers[col]);
+                    mapTileNum[col][row] = num;
+                    col++;
+                }
+                if (col == gp.maxScreenCol) {
+                    col = 0;
+                    row++;
+                }
+            }
+
+            br.close();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+    }
+
+    public void draw(Graphics2D g2) {
+
+        int col = 0;
+        int row = 0;
+        int x = 0;
+        int y = 0;
+
+        while (col < gp.maxScreenCol && row < gp.maxScreenRow) {
+            int tileNum = mapTileNum[col][row];
+            g2.drawImage(tile[tileNum].image, x, y, gp.tileSize, gp.tileSize, null);
+            x += gp.tileSize;
+            col++;
+
+            if (col == gp.maxScreenCol) {
+                col = 0;
+                x = 0;
+                row++;
+                y += gp.tileSize;
+            }
+        }
+    }
+
+}
